requirements.txt
python-dotenv
selenium
pytest
pytest-sugar